add typescript rename js -> ts update estlint rules
