+++ package.json
@@ -14,6 +14,7 @@
     "openai": "^4.11.0"
   },
   "devDependencies": {
+    "@typescript-eslint/parser": "^6.7.5",
     "dotenv": "^16.3.1",
     "eslint": "^8.50.0",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -21,7 +22,8 @@
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^5.0.0",
     "husky": "^8.0.3",
-    "prettier": "^3.0.3"
+    "prettier": "^3.0.3",
+    "typescript": "^5.2.2"
   },
   "homepage": "https://github.com/seriouslysean/monster-hunter-now-events#readme",
   "keywords": [
