+++ src/utils/chat-gpt.ts
