+++ tools/fetch-article.ts
@@ -1,8 +1,8 @@
 import { Command } from 'commander';
 
-import { version } from '../src/utils/config.js';
+import { version } from '../src/utils/config';
 
-import { getArticleByURL } from '../src/utils/article-utils.js';
+import { getArticleByURL } from '../src/utils/article-utils';
 
 // Invoke via `npm run fetch:article -- -u <url>`
 // HAS EVENTS:
